refactor: extract app stages into components
